+++ src/main/kotlin/com/github/jensim/megamanipulator/actions/git/GitWindow.kt
@@ -29,7 +29,8 @@ import com.intellij.ui.JBSplitter
 import com.intellij.ui.components.JBList
 import com.intellij.ui.components.JBScrollPane
 import com.intellij.ui.components.JBTextArea
-import com.intellij.ui.layout.panel
+import com.intellij.ui.dsl.builder.panel
+import com.intellij.ui.dsl.gridLayout.HorizontalAlign.RIGHT
 import java.awt.Color
 import java.awt.Dimension
 import java.awt.event.ActionEvent
@@ -44,6 +45,8 @@ private typealias DirResult = Pair<String, List<StepResult>>
 
 @SuppressWarnings("LongParameterList")
 class GitWindow(private val project: Project) : ToolWindowTab {
+
+    private val branchNamePattern = "^[a-zA-Z][a-zA-Z0-9/_-]*[a-zA-Z0-9]$"
 
     private val localRepoOperator: LocalRepoOperator by lazy { project.service() }
     private val settingsFileOperator: SettingsFileOperator by lazy { project.service() }
@@ -95,20 +98,17 @@ class GitWindow(private val project: Project) : ToolWindowTab {
 
     override val content: JComponent = panel {
         row {
-            buttonGroup {
-                component(btnListBranch)
-                component(btnSetBranch)
-                component(btnCommitAndPush)
-                component(btnJustPush)
-                component(btnCreatePRs)
-            }
-            component(btnCleanLocalClones)
-            right {
-                component(OnboardingButton(project, TabKey.tabTitleClones, OnboardingId.CLONES_TAB))
-            }
+            cell(btnListBranch)
+            cell(btnSetBranch)
+            cell(btnCommitAndPush)
+            cell(btnJustPush)
+            cell(btnCreatePRs)
+            cell(btnCleanLocalClones)
+            cell(OnboardingButton(project, TabKey.tabTitleClones, OnboardingId.CLONES_TAB))
+                .horizontalAlign(RIGHT)
         }
         row {
-            component(splitLeft)
+            cell(splitLeft)
         }
     }
 
@@ -134,18 +134,17 @@ class GitWindow(private val project: Project) : ToolWindowTab {
             refresh()
         }
         btnSetBranch.addActionListener {
-            val pattern = "^[a-zA-Z][a-zA-Z0-9/_-]*[a-zA-Z0-9]$"
             dialogGenerator.askForInput(
                 title = "Select branch name",
                 message = "This will NOT reset the repos to origin/default-branch first!!",
-                validationPattern = pattern,
+                validationPattern = branchNamePattern,
                 prefill = PrefillString.BRANCH,
                 focusComponent = btnSetBranch,
             ) { branch: String ->
-                if (branch.isBlank() || branch.isEmpty() || branch.contains(Regex(pattern))) {
+                if (branch.isBlank() || branch.isEmpty() || !branch.matches(Regex(branchNamePattern))) {
                     dialogGenerator.showConfirm(
                         title = "Bad branch name.",
-                        message = "$branch didnt match pattern $pattern",
+                        message = "$branch didnt match pattern $branchNamePattern",
                         yesText = "Ok",
                         noText = "Cancel",
                         focusComponent = btnSetBranch