+++ src/main/kotlin/com/github/jensim/megamanipulator/listener/OnStartListener.kt
@@ -11,6 +11,7 @@ import com.intellij.icons.AllIcons
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
+import com.intellij.openapi.wm.RegisterToolWindowTask
 import com.intellij.openapi.wm.ToolWindowAnchor
 import com.intellij.openapi.wm.ToolWindowManager
 import kotlinx.coroutines.runBlocking
@@ -23,11 +24,14 @@ class OnStartListener : StartupActivity {
 
     override fun runActivity(project: Project) {
         if (isMM(project)) {
-            ToolWindowManager.getInstance(project)?.let {
-                val toolWindow = it.registerToolWindow("Mega Manipulator") {
-                    this.icon = AllIcons.General.Modified
-                    this.anchor = ToolWindowAnchor.BOTTOM
-                }
+            ToolWindowManager.getInstance(project).let {
+                val toolWindow = it.registerToolWindow(
+                    RegisterToolWindowTask(
+                        id = "Mega Manipulator",
+                        anchor = ToolWindowAnchor.BOTTOM,
+                        icon = AllIcons.General.Modified,
+                    )
+                )
                 val factory = MyToolWindowFactory()
                 factory.createToolWindowContent(project, toolWindow)
                 toolWindow.show()