+++ src/main/kotlin/com/github/jensim/megamanipulator/ui/ClosePRDialogFactory.kt
@@ -3,7 +3,7 @@ package com.github.jensim.megamanipulator.ui
 import com.intellij.openapi.ui.popup.Balloon
 import com.intellij.openapi.ui.popup.JBPopupFactory
 import com.intellij.ui.components.JBCheckBox
-import com.intellij.ui.layout.panel
+import com.intellij.ui.dsl.builder.panel
 import javax.swing.JButton
 import javax.swing.JComponent
 
@@ -33,11 +33,11 @@ object ClosePRDialogFactory {
                         """.trimIndent()
                     )
                 }
-                row { component(pruneBranchBox) }
-                row { component(pruneForkBox) }
+                row { cell(pruneBranchBox) }
+                row { cell(pruneForkBox) }
                 row {
-                    component(okBtn)
-                    component(cancelBtn)
+                    cell(okBtn)
+                    cell(cancelBtn)
                 }
             }
             val popupFactory: JBPopupFactory = JBPopupFactory.getInstance()